midterm/requirements.txt
django-environ==0.9.0
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5